Second iteration for Windows. It fixes a number of problems that prevented dependent projects to work properly.
git-svn-id: svn+ssh://svn.cern.ch/reps/gaudi/Gaudi/branches/GAUDI/GAUDI_cmake@6347 53ecefc3-a64d-0410-9bdf-b8581fa3287e

+++ GaudiExamples/src/Selections/SelCreate.cpp
@@ -54,9 +54,9 @@ namespace Gaudi
           Gaudi::Examples::MyTrack* track = new Gaudi::Examples::MyTrack() ;
           
           // fill it with some "data"
-          track -> setPx ( gauss ()           ) ;
-          track -> setPy ( gauss ()           ) ;
-          track -> setPz ( gauss () + flat () ) ;
+          track -> setPx ( (float)gauss ()           ) ;
+          track -> setPy ( (float)gauss ()           ) ;
+          track -> setPz ( (float)gauss () + flat () ) ;
           
           // insert it into the container 
           tracks -> insert ( track ) ;
